Introduce exonum_build
Tiny crate to simplify protobuf code generation.
