+++ exonum/Cargo.toml
@@ -81,5 +81,4 @@ sodiumoxide-crypto = ["exonum_sodiumoxide"]
 with-serde = []
 
 [build-dependencies]
-protoc-rust = "=2.2.0"
-walkdir = "2.2.7"
+exonum_build = { version = "0.9.0", path = "../exonum_build" }
