+++ exonum/tests/explorer/blockchain/proto/mod.rs
@@ -20,6 +20,4 @@ pub use self::tests_transactions::{CreateWallet, Transfer};
 
 include!(concat!(env!("OUT_DIR"), "/exonum_tests_proto_mod.rs"));
 
-pub mod helpers {
-    pub use exonum::encoding::protobuf::helpers::*;
-}
+use exonum::encoding::protobuf::*;
