+++ exonum_build/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "exonum_build"
+version = "0.9.0"
+authors = ["The Exonum Team <exonum@bitfury.com>"]
+
+[dependencies]
+protoc-rust = "=2.2.0"
+walkdir = "2.2.7"
